Timer module
Created a simple, reusable timer module.

The start() method accepts a callback and an interval as arguments. The
callback is executed after every interval until the stop() method is
called.

The stop() method accepts an optional callback argument. It will clear
the current timer and execute the callback function, if present.